ai-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
langchain==0.0.340
langchain-google-genai==0.0.6
langchain-community==0.0.10
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.2

# Enhanced TTS with Murf.ai
murf==1.0.0
gTTS==2.3.2
pygame==2.5.2
